Lots of changes/fixes improvements
Now html gets logged out to a file with a static name, nothing is being printed in the terminal.
Numbers are formatted with a , and are still sortable in the tables.

Tables are now using the latest tablesorter version that allows number parsing.
The tables are now sorted by default by the % of changes and Num Rows respectively - to match the donut charts.
Trying to sort the data in JS in the header, seems to not work as intended.

Donut charts still buggy and don't correctly sort by size, but better than before.
Increased margin for someone being a minor author from 0.01 to 0.03.

Added exact timestamp of when the statistical information was gathered.
Formatted numbers with , for the metrics and responsibilities as well.

Changed eloc to "Effected Lines of Code"